[msh2vtu] overhaul, removed -k, log_level->log
in our OGS world there is no reason to keep the
gmsh data names and the wrong data types in the
meshes, which would happen if k was used.
Thus it is removed.

log_level argument was simplified to log (boolean)

main functionality now is based on pyvista which
simplifies it a lot

moved main functionality to separate file

add msh2vtu to top lvl API, mshes_from_gmsh to meshlib API