[meshlib] add scale function to ms, rm units
removes spatial_data_unit and spatial_output_unit
and time_unit. The more practical solution which
is now proposed is to scale the meshseries in the
beginning (spatially and temporally) and set the
output units in the plot setup, which are now only
anectotally - they do not to any unit conversion.